// apps/athena-server/src/engine/athena-config/enums/config-variables-group.enum.ts
export enum ConfigVariablesGroup {
  ServerConfig = 'server-config',
  RateLimitting = 'rate-limitting',
  StorageConfig = 'storage-config',
  GoogleAuth = 'google-auth',
  MicrosoftAuth = 'microsoft-auth',
  EmailSettings = 'email-settings',
  Logging = 'logging',
  Metering = 'metering',
  ExceptionHandler = 'exception-handler',
  Other = 'other',
  BillingConfig = 'billing-config',
  CaptchaConfig = 'captcha-config',
  CloudflareConfig = 'cloudflare-config',
  LLM = 'llm',
  ServerlessConfig = 'serverless-config',
  SSL = 'ssl',
  SupportChatConfig = 'support-chat-config',
  AnalyticsConfig = 'audit-config',
  TokensDuration = 'tokens-duration',
}